Add files via upload
The Attendance System using Face – Recognition is a replacement way method for the traditional way of marking attendance. The proposed system is python, tkinter based system supported with MySQL database. This system can be implemented on a single faculty system of a particular institute. This system is proposed to be based on biometrics .i.e. Face scans.
Authentication.
Since there is presence of biometrics, this system completely eliminates the chances of fake attendance which is a problem inherent with the traditional methods of attendance.